Added recommender file that takes query and songset args
The recommender.m file takes query and songset as arguments. It calculates the feature vectors for the query and considers the entire query a cluster. The feature vectors are then calculated for the songset. The plan is to implement EMD calcs for each song in the songset as a cluster compared with the query.

featurex.m was changed slightly, and it will continue to be updated, to pass on fully functioning feature data for EMD calculation.